+++ internal/repository/postgresql/pgbookmark/bookmark_integration_test.go
@@ -62,7 +62,7 @@ func TestBookmarkService(t *testing.T) {
 
 	fake := faker.New()
 
-	u := pgbase.GenerateFakeUser(t, &fake)
+	u := user.FakeUser(t, &fake)
 
 	if err := us.Add(t.Context(), u); err != nil {
 		t.Fatalf("failed to create user: %q", err)